Separate library code from examples.
