+++ README.md
@@ -1,0 +1,2 @@
+# farmOS-map Geotracing
+Functions and UI elements for tracing geographical features in real time, using device geolocation sensors.
