+++ examples/geotrace/sim.json
