+++ lib/Trace.js
