+++ lib/controlButton.js
