+++ lib/maths.js
